Add roundtrip tests for ProbabilisticGloveLayer
